# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Proyectos",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "https://backendchallenge-estoes.herokuapp.com/",
            "description": "[@OA\\Info(title=\"Challenge Backend Laravel\", version=\"1.0\")]"
        },
        {
            "url": "https://backendchallenge-estoes.herokuapp.com/",
            "description": "[@OA\\Info(title=\"Challenge Backend Laravel\", version=\"1.0\")]"
        },
        {
            "url": "https://backendchallenge-estoes.herokuapp.com/"
        }
    ],
    "paths": {
        "/api/login": {
            "post": {
                "summary": "Sign in",
                "description": "Login con email, password",
                "operationId": "3673edffcdc18c06ac5aef5e2219d774",
                "requestBody": {
                    "description": "Pasar credenciales de usuario",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "pao@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "12345678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "El email o password son incorrectos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Error, el email o password son incorrectos"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/register": {
            "post": {
                "summary": "Sign up",
                "description": "Registra un usuario",
                "operationId": "748a1acc92fb3e56f8f6ace55b5a9d24",
                "requestBody": {
                    "description": "Datos del usuario",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Marina"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "marina@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "12345678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "El email ya existe",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Error, el email ya existe"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/projects": {
            "get": {
                "summary": "Mostrar proyectos",
                "operationId": "8d4a13ef3fc773ff4031e515460919ea",
                "responses": {
                    "200": {
                        "description": "Mostrar todos los proyectos."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "post": {
                "summary": "Nuevo proyecto",
                "description": "Crea un nuevo proyecto",
                "operationId": "a4dd4c17bf5fc24c224357f03dcc848a",
                "requestBody": {
                    "description": "Datos del nuevo proyecto",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description",
                                    "status"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Proyecto laravel"
                                    },
                                    "description": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "This project runs with Laravel version 5.7"
                                    },
                                    "status": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "enabled"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/projects/{id}": {
            "get": {
                "summary": "Muestra un proyecto",
                "description": "Muestra un proyecto existente /id",
                "operationId": "af32aed47c19d517816e0b970d60cd5a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "example": 3
                    }
                ],
                "responses": {
                    "401": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "delete": {
                "summary": "Elimina un proyecto",
                "description": "Elimina un proyecto existente",
                "operationId": "187865e30ff6b72f1cea3037d9dd84fa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "401": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "patch": {
                "summary": "Actualiza proyecto",
                "description": "Actualiza un proyecto existente",
                "operationId": "79fb1fcf30a758547f944dc89bbb3120",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "example": 2
                    }
                ],
                "requestBody": {
                    "description": "Datos a modificar del proyecto",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Proyecto Nodejs"
                                    },
                                    "description": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "This project runs with NodeJS"
                                    },
                                    "status": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "enabled"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/buscar/?name={name}": {
            "get": {
                "summary": "Busca proyectos por nombres",
                "description": "Permite realizar una búsqueda por nombre de proyecto. ",
                "operationId": "b56c09979f8e87ffddc098701639c03a",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "required": true,
                        "example": "Proyecto"
                    }
                ],
                "responses": {
                    "401": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}